% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{Display message regarding environment variables}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{name of the library}

\item{pkgname}{name of the package}
}
\description{
Display message to user that environment variables were set for:
\itemize{
\item \code{SHAIDYMAPGEN} - Path to the included file ShaidyMapGen.jar
\item \code{NGCHMWIDGETPATH} - Path to the included file ngchmWidget-min.js
}
}
