% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Define environment variables}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{name of the library}

\item{pkgname}{name of the package}
}
\description{
Define environment variables:
\itemize{
\item \code{SHAIDYMAPGEN} - Path to the included file ShaidyMapGen.jar
\item \code{NGCHMWIDGETPATH} - Path to the included file ngchmWidget-min.js
}
}
